requirements.txt
Flask==3.0.3
pandas==2.2.0
openpyxl==3.1.2
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.1.2
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
gunicorn==21.2.0
